+++ client/reader/data/site-subscriptions/index.ts
@@ -2,4 +2,3 @@ export * from './cache';
 export * from './use-follow-mutations';
 export * from './use-follow-selectors';
 export * from './use-site-subscriptions';
-export * from './use-unsubscribe-with-undo';
