+++ client/reader/get-helpers.ts
@@ -272,3 +272,40 @@ export const getFollowerCount = ( feed: ReaderFeed, site: ReaderSite ): number |
 
 	return null;
 };
+
+interface GetSiteNameForSidebarArgs {
+	name?: string;
+	URL?: string;
+}
+
+const isFreeWpcomSubdomain = ( host = '' ): boolean => /\.wordpress\.com$/i.test( host );
+
+/**
+ * Label for a followed site: real title, else an `r/subreddit` handle, else the
+ * resolved domain. Untitled WordPress.com sites come back named after their free
+ * subdomain, so those fall through to the domain from `URL`.
+ */
+export function getReaderSidebarSiteName( site: GetSiteNameForSidebarArgs ): string {
+	const siteName = site.name ?? '';
+	// `name` may be URL-shaped, so normalize before the subdomain check.
+	const normalizedName = formatUrlForDisplay( siteName ) || siteName;
+
+	if ( siteName && ! isFreeWpcomSubdomain( normalizedName ) ) {
+		return siteName;
+	}
+
+	// A title-less subreddit reads best as its `r/name` (or `u/name`) handle, since
+	// every subreddit resolves to the same generic `reddit.com` domain. The host is
+	// anchored so only genuine `reddit.com` feeds qualify.
+	const reddit = site.URL?.match( /^https?:\/\/(?:[^/]+\.)?reddit\.com\/(r|user)\/([^/?#]+)/i );
+	if ( reddit ) {
+		return `${ reddit[ 1 ].toLowerCase() === 'user' ? 'u' : 'r' }/${ reddit[ 2 ] }`;
+	}
+
+	const siteDomain = site.URL ? getSiteDomain( { site: { URL: site.URL } } ) : undefined;
+	if ( siteDomain ) {
+		return siteDomain;
+	}
+
+	return siteName;
+}
